Put HTTP types in separate headers
