+++ src/base/http/request.h
@@ -1,6 +1,7 @@
 /*
  * Bittorrent Client using Qt and libtorrent.
- * Copyright (C) 2015  Vladimir Golovnev <glassez@yandex.ru>
+ * Copyright (C) 2014-2026  Vladimir Golovnev <glassez@yandex.ru>
+ * Copyright (C) 2018  Mike Tzou (Chocobo1)
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -28,29 +29,30 @@
 
 #pragma once
 
+#include <QByteArray>
+#include <QHash>
+#include <QList>
 #include <QString>
 
-#include "base/global.h"
-#include "constants.h"
-#include "header.h"
-#include "response.h"
+#include "headermap.h"
 
 namespace Http
 {
-    class ResponseBuilder
+    struct UploadedFile
     {
-    public:
-        void status(uint code = 200, const QString &text = u"OK"_s);
-        void setHeader(const Header &header);
-        void print(const QString &text, const QString &type = CONTENT_TYPE_HTML);
-        void print(const QByteArray &data, const QString &type = CONTENT_TYPE_HTML);
-        void clear();
+        QString filename;
+        QString type;  // MIME type
+        QByteArray data;
+    };
 
-        Response response() const;
-
-    private:
-        void print_impl(const QByteArray &data, const QString &type);
-
-        Response m_response;
+    struct Request
+    {
+        QString version;
+        QString method;
+        QString path;
+        HeaderMap headers;
+        QHash<QString, QByteArray> query;
+        QHash<QString, QString> posts;
+        QList<UploadedFile> files;
     };
 }
